requirements.txt
absl-py=2.1.0=pypi_0
annotated-types=0.6.0=pypi_0
anyio=4.3.0=pypi_0
astunparse=1.6.3=pypi_0
bzip2=1.0.8=he774522_0
ca-certificates=2024.2.2=h56e8100_0
cachetools=5.3.2=pypi_0
certifi=2024.2.2=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
colorama=0.4.6=pypi_0
cudatoolkit=11.2.2=h933977f_10
cudnn=8.1.0.77=h3e0f4f4_0
exceptiongroup=1.2.0=pypi_0
fastapi=0.109.2=pypi_0
filelock=3.9.0=pypi_0
flatbuffers=23.5.26=pypi_0
fsspec=2023.4.0=pypi_0
gast=0.4.0=pypi_0
google-auth=2.28.1=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.62.0=pypi_0
h11=0.14.0=pypi_0
h5py=3.10.0=pypi_0
httpcore=1.0.4=pypi_0
httpx=0.27.0=pypi_0
idna=3.6=pypi_0
jinja2=3.1.2=pypi_0
keras=2.10.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
libclang=16.0.6=pypi_0
libffi=3.4.4=hd77b12b_0
markdown=3.5.2=pypi_0
markupsafe=2.1.3=pypi_0
mpmath=1.3.0=pypi_0
networkx=3.2.1=pypi_0
nodeenv=1.8.0=pypi_0
numpy=1.26.3=pypi_0
oauthlib=3.2.2=pypi_0
opencv-python=4.9.0.80=pypi_0
openssl=3.0.13=h2bbff1b_0
opt-einsum=3.3.0=pypi_0
packaging=23.2=pypi_0
pandas=2.2.1=pypi_0
pillow=10.2.0=pypi_0
pip=23.3.1=py310haa95532_0
prisma=0.12.0=pypi_0
protobuf=3.19.6=pypi_0
pyasn1=0.5.1=pypi_0
pyasn1-modules=0.3.0=pypi_0
pydantic=2.6.2=pypi_0
pydantic-core=2.16.3=pypi_0
python=3.10.13=he1021f5_0
python-dateutil=2.8.2=pypi_0
python-dotenv=1.0.1=pypi_0
python-multipart=0.0.9=pypi_0
pytz=2024.1=pypi_0
requests=2.31.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
setuptools=68.2.2=py310haa95532_0
six=1.16.0=pypi_0
sniffio=1.3.0=pypi_0
sqlite=3.41.2=h2bbff1b_0
starlette=0.36.3=pypi_0
strenum=0.4.15=pypi_0
sympy=1.12=pypi_0
tensorboard=2.10.1=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow=2.10.1=pypi_0
tensorflow-estimator=2.10.0=pypi_0
tensorflow-io-gcs-filesystem=0.31.0=pypi_0
termcolor=2.4.0=pypi_0
tk=8.6.12=h2bbff1b_0
tomlkit=0.12.3=pypi_0
torch=2.2.1+cu121=pypi_0
torchaudio=2.2.1+cu121=pypi_0
torchvision=0.17.1+cu121=pypi_0
typing-extensions=4.8.0=pypi_0
tzdata=2024.1=pypi_0
urllib3=2.2.1=pypi_0
uvicorn=0.27.1=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
werkzeug=3.0.1=pypi_0
wheel=0.41.2=py310haa95532_0
wrapt=1.16.0=pypi_0
xz=5.4.5=h8cc25b3_0
zlib=1.2.13=h8cc25b3_0
